first cut of a folder restructure
