+++ AzureDailyCostCompare.Tests/TestHelper.cs
@@ -1,4 +1,5 @@
 ﻿
+using AzureDailyCostCompare.Application;
 using System.Text.Json;
 using System.Text.Json.Serialization;
