+++ AzureDailyCostCompare/Infrastructure/Interfaces/IAuthenticationService.cs
@@ -1,0 +1,8 @@
+﻿// Business logic interfaces and implementations
+
+namespace AzureDailyCostCompare.Infrastructure.Interfaces;
+
+public interface IAuthenticationService
+{
+    Task<string> GetAccessToken();
+}
