+++ AzureDailyCostCompare/Infrastructure/Interfaces/IBillingService.cs
@@ -1,0 +1,8 @@
+﻿// Business logic interfaces and implementations
+
+namespace AzureDailyCostCompare.Infrastructure.Interfaces;
+
+public interface IBillingService
+{
+    Task<string> GetBillingAccountIdAsync(string accessToken);
+}
